refactor manager login
